pyproject.toml
[tool.poetry]
name = "poetry-git-tags-plugin"
version = "1.1.0"
description = "Poetry plugin to create git tags when bumping versions"
authors = ["Mark Wibrow"]
license = "MIT"
readme = "README.md"


packages = [
    { include = "poetry_git_tags_plugin", from = "src" }
]

include = [
    { path = "tests", format = "sdist" }
]

[tool.poetry.dependencies]
python = "^3.6"
poetry = "^1.2.0a1"
GitPython = "^3.1.18"

[tool.poetry.dev-dependencies]
black = { version = "^21.4b1", python = "^3.6.2"}
isort = "^5.8.0"
pre-commit = { version = "^2.6", python = "^3.6.1" }
pytest = "^6.2.3"
pytest-mock = "^3.6.1"

[tool.poetry.plugins."poetry.application.plugin"]
export = "poetry_git_tags_plugin.plugins:VersionApplicationPlugin"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
